Move a few more interfaces into `secret-storage.ts`
